ComputerVision/FaceRecognation/requirements.txt
click
cmake
cvzone
face-recognition
face-recognition-models
numpy
opencv-python
pillow
pywin32
firebase_admin
